Update README.md
Added installation instructions for krnl and krnlc from the docs in the get started section. 
This will ensure clarity for people getting started installing with this crate.
In the docs more specific +nightly versions are used so i removed the version from the command. The new commands work fine on my macOS machine.